# src/docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
        "/veiculos": {
            "post": {
                "summary": "Cria um novo veículo",
                "description": "Insere veículo com marca, modelo, ano e placa.",
                "operationId": "criar_veiculo_veiculos_post",
                "parameters": [
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VeiculoCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "Lista todos os veículos",
                "description": "Lista a quantidade de veículos cadastrados no sistema.",
                "operationId": "listar_veiculos_veiculos_get",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/veiculos/filtro": {
            "get": {
                "summary": "Filtra veículos",
                "description": "Filtra veículos por ID, marca, modelo, ano ou placa.",
                "operationId": "filtrar_veiculos_veiculos_filtro_get",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Id"
                        }
                    },
                    {
                        "name": "marca",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Marca"
                        }
                    },
                    {
                        "name": "modelo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Modelo"
                        }
                    },
                    {
                        "name": "ano",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Ano"
                        }
                    },
                    {
                        "name": "placa",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Placa"
                        }
                    },
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/veiculos/{veiculo_id}": {
            "put": {
                "summary": "Atualiza um veículo",
                "description": "Atualiza os dados de um veículo existente.",
                "operationId": "atualizar_veiculo_veiculos__veiculo_id__put",
                "parameters": [
                    {
                        "name": "veiculo_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Veiculo Id"
                        }
                    },
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VeiculoUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Deleta um veículo",
                "description": "Remove um veículo do sistema pelo ID.",
                "operationId": "deletar_veiculo_veiculos__veiculo_id__delete",
                "parameters": [
                    {
                        "name": "veiculo_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Veiculo Id"
                        }
                    },
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
            "VeiculoCreate": {
                "properties": {
                    "marca": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 2,
                        "title": "Marca"
                    },
                    "modelo": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 1,
                        "title": "Modelo"
                    },
                    "ano": {
                        "type": "integer",
                        "maximum": 2100.0,
                        "minimum": 1886.0,
                        "title": "Ano"
                    },
                    "placa": {
                        "type": "string",
                        "maxLength": 8,
                        "minLength": 7,
                        "title": "Placa"
                    }
                },
                "type": "object",
                "required": [
                    "marca",
                    "modelo",
                    "ano",
                    "placa"
                ],
                "title": "VeiculoCreate"
            },
            "VeiculoUpdate": {
                "properties": {
                    "marca": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 50,
                                "minLength": 2
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Marca"
                    },
                    "modelo": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 50,
                                "minLength": 1
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Modelo"
                    },
                    "ano": {
                        "anyOf": [
                            {
                                "type": "integer",
                                "maximum": 2100.0,
                                "minimum": 1886.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ano"
                    },
                    "placa": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 8,
                                "minLength": 7
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Placa"
                    }
                },
                "type": "object",
                "title": "VeiculoUpdate"
            }
        }
    }
}